feat: add wordbook library screen
